+++ Sources/BadgeIcon/BadgeIcon.swift
@@ -17,20 +17,25 @@ public struct BadgeIcon: View {
     /**
      Create a badge icon.
      
+     The size-based parameters, like `iconOffset`, should be
+     expressed as a `0.0-1.0` percentage of the icon size to
+     let them scale well with the icon size. For instance, a
+     `0.50` `badgeCornerRadius` will make the badge circular.
+     
      - Parameters:
        - icon: The image to use.
        - iconColor: The icon color, by default `semi-black` or `white`.
        - iconColorScheme: The icon color scheme, by default `nil`.
        - iconFill: The icon fill mode, by default `true`.
        - iconGradient: Whether or not to add a gradient to the icon color, by default `true`.
-       - iconOffset: The icon offset, by default `.zero`.
-       - iconPadding: The icon padding, by default calculated at runtime.
+       - iconOffset: The icon offset, by default `0` of the icon size.
+       - iconPadding: The icon padding, by default `0.15` of the icon size.
        - iconRenderingMode: The icon symbol rendering mode, by default `.monochrome`.
        - badgeColor: The badge color, by default `.white`.
-       - badgeCornerRadius: The badge corner radius, by default calculated at runtime.
+       - badgeCornerRadius: The badge corner radius, by default `0.3` of the icon size.
        - badgeGradient: Whether or not to add a gradient to the icon color, by default `true`.
        - badgeStrokeColor: The badge stroke color, if any.
-       - badgeStrokeWidth: The badge stroke width, by default `1`.
+       - badgeStrokeWidth: The badge stroke width, by default `0.3` of the icon size.
      */
     public init(
         icon: Image,
@@ -39,21 +44,19 @@ public struct BadgeIcon: View {
         iconFill: Bool = true,
         iconGradient: Bool = true,
         iconOffset: CGPoint = .zero,
-        iconPadding: Double? = nil,
+        iconPadding: Double = 0.15,
         iconRenderingMode: SymbolRenderingMode = .monochrome,
         badgeColor: Color = .white,
-        badgeCornerRadius: Double? = nil,
+        badgeCornerRadius: Double = 0.3,
         badgeGradient: Bool = true,
         badgeStrokeColor: Color? = nil,
-        badgeStrokeWidth: Double = 1
+        badgeStrokeWidth: Double = 0.001
     ) {
         let whiteBadge = badgeColor == .white
         let whiteBadgeStroke = Color.hex(0xe7e7e7)
         let whiteBadgeIconColor = Color.black.opacity(0.8)
-        let defaultIconColor = Color.white
-        let defaultStrokeColor = Color.clear
-        let fallbackIconColor = whiteBadge ? whiteBadgeIconColor : defaultIconColor
-        let fallbackStroke = whiteBadge ? whiteBadgeStroke : defaultStrokeColor
+        let fallbackIconColor = whiteBadge ? whiteBadgeIconColor : .white
+        let fallbackStroke = whiteBadge ? whiteBadgeStroke : .clear
         
         self.icon = icon
         self.iconColor = iconColor ?? fallbackIconColor
@@ -76,10 +79,10 @@ public struct BadgeIcon: View {
     public var iconGradient: Bool
     public var iconFill: Bool
     public var iconOffset: CGPoint
-    public var iconPadding: Double?
+    public var iconPadding: Double
     public var iconRenderingMode: SymbolRenderingMode
     public var badgeColor: Color
-    public var badgeCornerRadius: Double?
+    public var badgeCornerRadius: Double
     public var badgeGradient: Bool
     public var badgeStrokeColor: Color
     public var badgeStrokeWidth: Double
@@ -94,7 +97,7 @@ public struct BadgeIcon: View {
                     .cornerRadius(cornerRadius(for: geo))
                     .overlay(
                         RoundedRectangle(cornerRadius: cornerRadius(for: geo))
-                            .stroke(badgeStrokeColor, lineWidth: badgeStrokeWidth)
+                            .stroke(badgeStrokeColor, lineWidth: strokeWidth(for: geo))
                     )
                     .aspectRatio(1, contentMode: .fit)
                     .overlay(
@@ -104,7 +107,7 @@ public struct BadgeIcon: View {
                             .symbolRenderingMode(iconRenderingMode)
                             .symbolVariant(iconFill ? .fill : .none)
                             .padding(iconPadding(for: geo))
-                            .offset(x: iconOffset.x, y: iconOffset.y)
+                            .offset(iconOffset(for: geo))
                             .foreground(iconColor, gradient: iconGradient)
                     )
             }
@@ -113,14 +116,34 @@ public struct BadgeIcon: View {
     }
 }
 
+private extension View {
+    
+    func offset(_ point: CGPoint) -> some View {
+        self.offset(x: point.x, y: point.y)
+    }
+}
+
 extension BadgeIcon {
     
     func cornerRadius(for geo: GeometryProxy) -> Double {
-        badgeCornerRadius ?? 0.3 * geo.size.width
+        badgeCornerRadius * geo.size.width
+    }
+    
+    func iconOffset(for geo: GeometryProxy) -> CGPoint {
+        let width = geo.size.width
+        let offset = iconOffset
+        return CGPoint(
+            x: width * offset.x,
+            y: width * offset.y
+        )
     }
     
     func iconPadding(for geo: GeometryProxy) -> Double {
-        badgeCornerRadius ?? 0.17 * geo.size.width
+        iconPadding * geo.size.width
+    }
+    
+    func strokeWidth(for geo: GeometryProxy) -> Double {
+        max(badgeStrokeWidth * geo.size.width, 1)
     }
 }
 
@@ -168,34 +191,34 @@ private extension View {
             iconColorScheme: nil,
             iconFill: true,
             iconGradient: false,
-            iconOffset: .init(x: -45, y: -50),
-            iconPadding: 2,
+            iconOffset: .init(x: -0.4, y: -0.4),
+            iconPadding: 0.2,
             iconRenderingMode: .monochrome,
             badgeColor: .indigo,
-            badgeCornerRadius: 20,
+            badgeCornerRadius: 0.4,
             badgeGradient: true,
             badgeStrokeColor: .teal,
-            badgeStrokeWidth: 10
+            badgeStrokeWidth: 0.05
         )
-        .bold()
-        .shadow(radius: 10)
+        
         BadgeIcon(
             icon: .symbol("checkmark"),
             iconColor: .green,
             iconColorScheme: nil,
             iconFill: true,
             iconGradient: true,
-            iconOffset: .init(x: 45, y: -50),
-            iconPadding: 2,
+            iconOffset: .init(x: 0.25, y: -0.25),
+            iconPadding: 0.1,
             iconRenderingMode: .monochrome,
             badgeColor: .red,
-            badgeCornerRadius: 20,
+            badgeCornerRadius: 0.3,
             badgeGradient: true,
             badgeStrokeColor: .blue,
-            badgeStrokeWidth: 20
+            badgeStrokeWidth: 0.05
         )
         .bold()
-        .shadow(radius: 10)
+        
+        BadgeIcon.calendar
     }
     .padding(100)
 }